adding more tests

+++ src/main/java/net/seninp/jmotif/sax/motif/MotifRecord.java
@@ -40,7 +40,7 @@ public class MotifRecord {
    * 
    * @return the motif location.
    */
-  public double getLocation() {
+  public int getLocation() {
     return location;
   }
